[#6758] move validation logic to Sormas2SormasDataValidator
