+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sormastosormas/entities/caze/ReceivedCaseProcessor.java
@@ -20,34 +20,25 @@ import static de.symeda.sormas.backend.sormastosormas.ValidationHelper.buildCont
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import javax.ejb.EJB;
 import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
 
-import de.symeda.sormas.api.EntityDto;
 import de.symeda.sormas.api.caze.CaseDataDto;
-import de.symeda.sormas.api.caze.maternalhistory.MaternalHistoryDto;
-import de.symeda.sormas.api.contact.ContactDto;
-import de.symeda.sormas.api.infrastructure.facility.FacilityType;
 import de.symeda.sormas.api.i18n.Captions;
 import de.symeda.sormas.api.person.PersonDto;
 import de.symeda.sormas.api.sormastosormas.SormasToSormasOriginInfoDto;
 import de.symeda.sormas.api.sormastosormas.SormasToSormasSampleDto;
 import de.symeda.sormas.api.sormastosormas.validation.SormasToSormasValidationException;
-import de.symeda.sormas.api.sormastosormas.validation.ValidationErrorMessage;
 import de.symeda.sormas.api.sormastosormas.validation.ValidationErrors;
 import de.symeda.sormas.api.sormastosormas.caze.SormasToSormasCaseDto;
 import de.symeda.sormas.api.sormastosormas.sharerequest.SormasToSormasCasePreview;
 import de.symeda.sormas.api.sormastosormas.sharerequest.SormasToSormasContactPreview;
-import de.symeda.sormas.api.utils.DataHelper;
 import de.symeda.sormas.backend.contact.ContactFacadeEjb.ContactFacadeEjbLocal;
 import de.symeda.sormas.backend.sormastosormas.data.infra.InfrastructureValidator;
 import de.symeda.sormas.backend.sormastosormas.data.received.ReceivedDataProcessor;
-import de.symeda.sormas.backend.sormastosormas.data.received.ReceivedDataProcessorHelper;
+import de.symeda.sormas.backend.sormastosormas.data.Sormas2SormasDataValidator;
 
 @Stateless
 @LocalBean
@@ -55,7 +46,7 @@ public class ReceivedCaseProcessor
 	implements ReceivedDataProcessor<CaseDataDto, SormasToSormasCaseDto, ProcessedCaseData, SormasToSormasCasePreview> {
 
 	@EJB
-	private ReceivedDataProcessorHelper dataProcessorHelper;
+	private Sormas2SormasDataValidator dataValidator;
 	@EJB
 	private ContactFacadeEjbLocal contactFacade;
 	@EJB
@@ -74,10 +65,10 @@ public class ReceivedCaseProcessor
 
 		ValidationErrors caseValidationErrors = new ValidationErrors();
 
-		ValidationErrors originInfoErrorsErrors = dataProcessorHelper.processOriginInfo(originInfo, Captions.CaseData);
+		ValidationErrors originInfoErrorsErrors = dataValidator.validateOriginInfo(originInfo, Captions.CaseData);
 		caseValidationErrors.addAll(originInfoErrorsErrors);
 
-		ValidationErrors caseDataErrors = processCaseData(caze, person, existingCaseData);
+		ValidationErrors caseDataErrors = dataValidator.validateCaseData(caze, person, existingCaseData);
 		caseValidationErrors.addAll(caseDataErrors);
 
 		if (caseValidationErrors.hasError()) {
@@ -85,12 +76,12 @@ public class ReceivedCaseProcessor
 		}
 
 		if (associatedContacts != null && associatedContacts.size() > 0) {
-			List<ValidationErrors> contactValidationErrors = processAssociatedContacts(associatedContacts);
+			List<ValidationErrors> contactValidationErrors = dataValidator.validateAssociatedContacts(associatedContacts);
 			validationErrors.addAll(contactValidationErrors);
 		}
 
 		if (samples != null && samples.size() > 0) {
-			List<ValidationErrors> sampleErrors = dataProcessorHelper.processSamples(samples);
+			List<ValidationErrors> sampleErrors = dataValidator.validateSamples(samples);
 			validationErrors.addAll(sampleErrors);
 		}
 
@@ -105,175 +96,24 @@ public class ReceivedCaseProcessor
 	public SormasToSormasCasePreview processReceivedPreview(SormasToSormasCasePreview preview) throws SormasToSormasValidationException {
 		List<ValidationErrors> validationErrors = new ArrayList<>();
 
-		ValidationErrors caseValidationErrors = new ValidationErrors();
-
-		DataHelper.Pair<InfrastructureValidator.InfrastructureData, List<ValidationErrorMessage>> infrastructureAndErrors =
-			infraValidator.validateInfrastructure(
-				null,
-				null,
-				null,
-				preview.getRegion(),
-				preview.getDistrict(),
-				preview.getCommunity(),
-				preview.getFacilityType(),
-				preview.getHealthFacility(),
-				preview.getHealthFacilityDetails(),
-				preview.getPointOfEntry(),
-				preview.getPointOfEntryDetails());
-
-		infraValidator.handleInfraStructure(infrastructureAndErrors, Captions.CaseData, caseValidationErrors, infrastructureData -> {
-			preview.setRegion(infrastructureData.getRegion());
-			preview.setDistrict(infrastructureData.getDistrict());
-			preview.setCommunity(infrastructureData.getCommunity());
-			preview.setHealthFacility(infrastructureData.getFacility());
-			preview.setHealthFacilityDetails(infrastructureData.getFacilityDetails());
-			preview.setPointOfEntry(infrastructureData.getPointOfEntry());
-			preview.setPointOfEntryDetails(infrastructureData.getPointOfEntryDetails());
-		});
+		ValidationErrors caseValidationErrors = dataValidator.validateCasePreview(preview);
 
 		if (caseValidationErrors.hasError()) {
 			validationErrors.add(new ValidationErrors(buildCaseValidationGroupName(preview), caseValidationErrors));
 		}
 
-		ValidationErrors personValidationErrors = dataProcessorHelper.processPersonPreview(preview.getPerson());
+		ValidationErrors personValidationErrors = dataValidator.validatePersonPreview(preview.getPerson());
 		caseValidationErrors.addAll(personValidationErrors);
 
 		List<SormasToSormasContactPreview> contacts = preview.getContacts();
 		if (contacts != null && contacts.size() > 0) {
-			List<ValidationErrors> contactValidationErrors = processContactPreviews(contacts);
+			List<ValidationErrors> contactValidationErrors = dataValidator.validateContactPreviews(contacts);
 			validationErrors.addAll(contactValidationErrors);
 		}
 
 		if (validationErrors.size() > 0) {
 			throw new SormasToSormasValidationException(validationErrors);
 		}
-
 		return preview;
-	}
-
-	private ValidationErrors processCaseData(CaseDataDto caze, PersonDto person, CaseDataDto existingCaseData) {
-		ValidationErrors caseValidationErrors = new ValidationErrors();
-
-		ValidationErrors personValidationErrors = dataProcessorHelper.processPerson(person);
-		caseValidationErrors.addAll(personValidationErrors);
-
-		caze.setPerson(person.toReference());
-		dataProcessorHelper.updateReportingUser(caze, existingCaseData);
-
-		DataHelper.Pair<InfrastructureValidator.InfrastructureData, List<ValidationErrorMessage>> infrastructureAndErrors =
-			infraValidator.validateInfrastructure(caze);
-
-		infraValidator.handleInfraStructure(infrastructureAndErrors, Captions.CaseData, caseValidationErrors, infrastructureData -> {
-			caze.setResponsibleRegion(infrastructureData.getResponsibleRegion());
-			caze.setResponsibleDistrict(infrastructureData.getResponsibleDistrict());
-			caze.setResponsibleCommunity(infrastructureData.getResponsibleCommunity());
-			caze.setRegion(infrastructureData.getRegion());
-			caze.setDistrict(infrastructureData.getDistrict());
-			caze.setCommunity(infrastructureData.getCommunity());
-			caze.setHealthFacility(infrastructureData.getFacility());
-			caze.setHealthFacilityDetails(infrastructureData.getFacilityDetails());
-			caze.setPointOfEntry(infrastructureData.getPointOfEntry());
-			caze.setPointOfEntryDetails(infrastructureData.getPointOfEntryDetails());
-		});
-
-		ValidationErrors embeddedObjectErrors = processEmbeddedObjects(caze);
-		caseValidationErrors.addAll(embeddedObjectErrors);
-
-		return caseValidationErrors;
-	}
-
-	private ValidationErrors processEmbeddedObjects(CaseDataDto caze) {
-		ValidationErrors validationErrors = new ValidationErrors();
-
-		if (caze.getHospitalization() != null) {
-
-			caze.getHospitalization().getPreviousHospitalizations().forEach(ph -> {
-
-				DataHelper.Pair<InfrastructureValidator.InfrastructureData, List<ValidationErrorMessage>> phInfrastructureAndErrors =
-					infraValidator.validateInfrastructure(
-						null,
-						null,
-						null,
-						ph.getRegion(),
-						ph.getDistrict(),
-						ph.getCommunity(),
-						FacilityType.HOSPITAL,
-						ph.getHealthFacility(),
-						ph.getHealthFacilityDetails(),
-						null,
-						null);
-
-				infraValidator.handleInfraStructure(
-					phInfrastructureAndErrors,
-					Captions.CaseHospitalization_previousHospitalizations,
-					validationErrors,
-					(phInfrastructure) -> {
-						ph.setRegion(phInfrastructure.getRegion());
-						ph.setDistrict(phInfrastructure.getDistrict());
-						ph.setCommunity(phInfrastructure.getCommunity());
-						ph.setHealthFacility(phInfrastructure.getFacility());
-						ph.setHealthFacilityDetails(phInfrastructure.getFacilityDetails());
-					});
-			});
-		}
-
-		MaternalHistoryDto maternalHistory = caze.getMaternalHistory();
-		if (maternalHistory != null) {
-
-			DataHelper.Pair<InfrastructureValidator.InfrastructureData, List<ValidationErrorMessage>> rashExposureInfrastructureAndErrors =
-				infraValidator.validateInfrastructure(
-					maternalHistory.getRashExposureRegion(),
-					maternalHistory.getRashExposureDistrict(),
-					maternalHistory.getRashExposureCommunity());
-
-			infraValidator.handleInfraStructure(
-				rashExposureInfrastructureAndErrors,
-				Captions.MaternalHistory_rashExposure,
-				validationErrors,
-				(rashExposureInfrastructure) -> {
-					maternalHistory.setRashExposureRegion(rashExposureInfrastructure.getRegion());
-					maternalHistory.setRashExposureDistrict(rashExposureInfrastructure.getDistrict());
-					maternalHistory.setRashExposureCommunity(rashExposureInfrastructure.getCommunity());
-				});
-		}
-
-		dataProcessorHelper.processEpiData(caze.getEpiData(), validationErrors);
-
-		return validationErrors;
-	}
-
-	private List<ValidationErrors> processAssociatedContacts(List<SormasToSormasCaseDto.AssociatedContactDto> associatedContacts) {
-		List<ValidationErrors> validationErrors = new ArrayList<>();
-
-		Map<String, ContactDto> existingContactsMap =
-			contactFacade.getByUuids(associatedContacts.stream().map(c -> c.getContact().getUuid()).collect(Collectors.toList()))
-				.stream()
-				.collect(Collectors.toMap(EntityDto::getUuid, Function.identity()));
-
-		for (SormasToSormasCaseDto.AssociatedContactDto associatedContact : associatedContacts) {
-			ContactDto contact = associatedContact.getContact();
-			ValidationErrors contactErrors =
-				dataProcessorHelper.processContactData(contact, associatedContact.getPerson(), existingContactsMap.get(contact.getUuid()));
-
-			if (contactErrors.hasError()) {
-				validationErrors.add(new ValidationErrors(buildContactValidationGroupName(contact), contactErrors));
-			}
-		}
-
-		return validationErrors;
-	}
-
-	private List<ValidationErrors> processContactPreviews(List<SormasToSormasContactPreview> contacts) {
-		List<ValidationErrors> validationErrors = new ArrayList<>();
-
-		for (SormasToSormasContactPreview contact : contacts) {
-			ValidationErrors contactErrors = dataProcessorHelper.processContactPreview(contact);
-
-			if (contactErrors.hasError()) {
-				validationErrors.add(new ValidationErrors(buildContactValidationGroupName(contact), contactErrors));
-			}
-		}
-
-		return validationErrors;
 	}
 }
